It runs lambdas, but it's a complete mess
The implementation is a complete disgrace so far, but it runs
a simple lambda and correctly creates closures over arguments.
I've attempted to refactor the solution a couple of times while
in the TDD loop, but just the current functionally doesn't offer
me enough information to choose a good structure, so I will grow
the solution a bit more before attempting to shape it properly.